* Many files: Changes to avoid gcc warnings: Add ATTRIBUTE_UNUSED 	as appropriate.  Fill in structure initializations.  Add variable 	initializations.  Add casts. 	* dwarf2dbg.c (print_stats): Change i to size_t. 	* listing.c (listing_listing): Change list_line to unsigned int.

+++ gas/app.c
@@ -267,7 +267,7 @@ app_pop (arg)
     saved_input = NULL;
   else
     {
-      assert (saved->saved_input_len <= sizeof input_buffer);
+      assert (saved->saved_input_len <= (int) (sizeof input_buffer));
       memcpy (input_buffer, saved->saved_input, saved->saved_input_len);
       saved_input = input_buffer;
       saved_input_len = saved->saved_input_len;
